Add LoggingArgs
