+++ Cargo.toml
@@ -10,6 +10,7 @@ homepage = "https://github.com/jelmer/janitor"
 build = "build.rs"
 
 [dependencies]
+clap = { version = "4", features = ["derive"] }
 env_logger = "0.10.0"
 log = "0.4.19"
 protobuf = "3"
